Fix running with release build.
The unsafe code to call ioctl()s was corrupting memory because
allocations weren't happening. The cause was lack of interior mutability
in the calls to libc::ioctl() inside the unsafe code. Even though the dq
variable was mutable the binding when calling libc::ioctl() wasn't, and
that has undefined behavior. With no optimizations in the dev build
everything worked but then failed with all optimizations of the release
build.