Handle device stats enabled or not.
Device stats are going to be enabled only when running as root since
most of the data needs those privileges to be read. Running as a normal
user shows the accesible DRM clients data. This way the app doesn't
abort with "Permission denied" errors (or others) due lack of
privileges.