Display as much information as possible as non-root user.
This reverts a big part of commit d0861e2, and includes additional
changes to bring back more information for non-root users, including:
 - open render node to issue ioctl()s to amdgpu driver
 - create DrmDriver when all device nodes are known and stop getting
   freqs, mem info and power too early

Even with that, the user needs to be part of the right groups in their
Linux distribution to be able to open either legacy/primary DRM device
nodes (card*) or render ones (renderD*). Those groups on Fedora are
video and render.