app: update stats at fixed intervals and UI whenever needed
Disconnects the stats updates from UI ones. The stats will be updated at
default value or ms interval from command line option. The UI is updated
when user events are processed or when stats are updated.

This removes the need for a more complex mainloop. If at any point a
stat needs to be updated and/or processed at different interval than the
main app, that needs to be handled within the data source (likely through
another thread or process). Dropping the mainloop TODO item.