lib: support for any number of freqs/limits.
Data layer is adjusted for reading any number of frequency data from
drivers. The implementation is done in a generic way for i915 and Xe
KMDs on all present GTs, but amdgpu is reporting just GFX frequencies
for now.

In this commit, the TUI is kept displaying freqs data just for the
GT/GFX engines (the first one reported by the drivers).