plotter: reduce and improve device charts plotted
Limit charts to be per device and just 4 type: memory info, engines
usage, frequencies, and power usage.

Command-line options can select which device to plot charts for
(default: all of them) and which charts (default: all the 4 charts).

Each chart is saved into a separate SVG file with the prefix provided in
the command line switch for the plot sub-command.

Closes: #5